requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.1.0
djangorestframework==3.7.3
graphene==2.0.1
graphene-django==2.0.0
graphene-django-extras==0.2.2
graphql-core==2.0
graphql-relay==0.4.5
idna==2.6
iso8601==0.1.12
model-mommy==1.4.0
olefile==0.44
Pillow==4.3.0
promise==2.1
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
Rx==1.6.0
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.2
urllib3==1.22
